👩‍💻‍ Add package script test:update-fixtures
